vimhelp: Add offline light/dark mode support
Make light/dark mode completely offline, instead of relying on server.
Currently the logic is already almost completely client-side via cookie,
but it sends the requested mode to the server, which pre-generates the
correct version to the client. Instead, we just have a simple JavaScript
to set it on launch. It does mean this won't work on web browsers that
turned off JS, but that's not a big deal.

Also fixed up the way cookies work as it has some special casing to make
sure neo.vimhelp.org and vimhelp.org could share the same cookies, but
that's irrelevant to MacVim's use cases and should be removed.