vimhelp: Better header/footer and metadata on the web page
Add support for specifying the version and commit on the command line.

The footer will now show something that's more specific to MacVim, while
still leaving attribution for vimhelp with a link. It will show the
timestamp for generation time and the commit the docs were generated
from, so that a viewer can know easily whether the docs are up-to-date,
which it should be unless CI broke.

The navigation menu on top and bottom now also has a "MacVim reference"
to quickly navigate to `gui_mac.txt`, since that's likely the reason for
someone browsing this to begin with (instead of vanilla Vim's
documentation).