Add GitHub CI for deployment and corresponding scripts
Set up GitHub Action CI to deploy page. Currently using the new GitHub
feature to directly publish an artifact as a deployment instead of
needing to go through a separate gh-pages branch. This makes sense for
this repository because everything is generated by the vimhelp tool
anyway and we never need to directly commit anything to a branch. By
using an artifact based approach we won't have to worry about the Git
repository getting too big (e.g. Neovim actually has a manual history
pruning step to prune their gh-pages branch for this reason).

Added some scripts to manually parse MacVim's version info from the
source files so we can display that in the main page.

Currently, the CI is going to run every night, which is not ideal. We
should find a way to cache the last version we generated, and do not
re-regenerate it unless this repo itself has changed, or MacVim has been
updated.