vimhelp: Add redirect page which allows us to serve a permanent link
This way, even if a tag changes location later, the redirect will always
be able to find it and redirect to the correct location. Only issue is
that it does require JavaScript to work.

The way it works is by embedding the tags.json directly in the page, and
then the JS can just simply do a check to see if the tag match.