vimhelp: Add offline tag search ability using tags.json
Add back the search box when using static page generation. This works by
querying tags.json and then performing a local search instead of relying
on the server code in tagsearch.py. The matching logic is ported over to
be the same though. Because we only have ~10k tags, linear search is
fast enough that we don't need to implement a binary search (since JS
doesn't come with a native binary search algorithm). Otherwise, for
performance, a indexed array using the initial character of the tag may
work better than implementing a binary search.