Prevent deploying the same docs when nothing has changed
Use GitHub Action caching mechanism to store the last built versions,
keyed using both this repo's and MacVim's commit. This prevents us from
unnecessarily re-generating the page every day and help with browser
caching behaviors. We still build the page regardless as it finishes
quickly and good for sanity checking that nothing broke.

We do allow manual workflow dispatch to always deploy, since that's a
manual action that indicates we want to do so.